Initial commit to development branch, focused on getting a simple version of the algorithm to run on pi.
This .slx simulink project should run when loaded. It reads in the star_field1.mat and star_table1.mat files, and simulates a spacecraft spinnning at a constant rate. It crops the reference photo of the starfield to generate synthetic images to calculate attitude from, and generates synthetic sun sensor images. As it runs, it prints out the current attitude, as well as the attitudes determined from these two sensors. The sun sensor works well when the sun is in the field of view (sun is at phi=0), with a coarse accuracy. The star tracker is more accurate, but is consistently wrong at certain attitudes.